MSN-Scanner/requirements.txt
python-nmap
fpdf
requests
